fixes data module

+++ data/src/main/java/com/emikhalets/data/repository/DatabaseRepositoryImpl.kt
@@ -1,14 +1,16 @@
 package com.emikhalets.data.repository
 
-import com.emikhalets.core.CATEGORY_EXPENSE_ID
-import com.emikhalets.core.CATEGORY_INCOME_ID
 import com.emikhalets.core.CODE_CATEGORY_EXISTED
-import com.emikhalets.core.CODE_DELETE_CURRENCY_TRANSACTIONS
-import com.emikhalets.core.CURRENCY_ID
+import com.emikhalets.core.CODE_DELETE_CURRENCY_WITH_TRANSACTIONS
+import com.emikhalets.core.DEFAULT_CATEGORY_EXPENSE_ID
+import com.emikhalets.core.DEFAULT_CATEGORY_INCOME_ID
+import com.emikhalets.core.Prefs
 import com.emikhalets.core.UiString
-import com.emikhalets.core.WALLET_ID
-import com.emikhalets.data.CATEGORY_EXISTED
-import com.emikhalets.data.DELETE_CURRENCY_TRANSACTIONS
+import com.emikhalets.data.ERROR_CATEGORY_EXISTED
+import com.emikhalets.data.ERROR_CURRENCY_EXISTED
+import com.emikhalets.data.ERROR_DELETE_CURRENCY_WITH_TRANSACTIONS
+import com.emikhalets.data.ERROR_WALLET_EXISTED
+import com.emikhalets.data.R
 import com.emikhalets.data.database.dao.CategoriesDao
 import com.emikhalets.data.database.dao.CurrenciesDao
 import com.emikhalets.data.database.dao.TransactionsDao
@@ -20,6 +22,7 @@ import com.emikhalets.domain.entity.ResultWrapper
 import com.emikhalets.domain.entity.TransactionEntity
 import com.emikhalets.domain.entity.TransactionType
 import com.emikhalets.domain.entity.WalletEntity
+import com.emikhalets.domain.entity.complex.ComplexTransactionEntity
 import com.emikhalets.domain.entity.complex.ComplexWalletEntity
 import com.emikhalets.domain.repository.DatabaseRepository
 import kotlinx.coroutines.flow.Flow
@@ -32,6 +35,7 @@ class DatabaseRepositoryImpl @Inject constructor(
     private val walletsDao: WalletsDao,
     private val currenciesDao: CurrenciesDao,
     private val mapper: DatabaseMapper,
+    private val prefs: Prefs,
 ) : DatabaseRepository {
 
     private suspend fun <T : Any> execute(block: suspend () -> T): ResultWrapper<T> {
@@ -40,16 +44,23 @@ class DatabaseRepositoryImpl @Inject constructor(
         } catch (e: Exception) {
             e.printStackTrace()
             when (e.message) {
-                DELETE_CURRENCY_TRANSACTIONS -> {
-                    ResultWrapper.Error(UiString.create(), CODE_DELETE_CURRENCY_TRANSACTIONS)
+                ERROR_DELETE_CURRENCY_WITH_TRANSACTIONS -> {
+                    ResultWrapper.Error(UiString.create(), CODE_DELETE_CURRENCY_WITH_TRANSACTIONS)
                 }
-                CATEGORY_EXISTED -> {
+                ERROR_CATEGORY_EXISTED -> {
                     ResultWrapper.Error(UiString.create(), CODE_CATEGORY_EXISTED)
                 }
                 else -> {
-                    ResultWrapper.Error(UiString.create(e.message))
+                    ResultWrapper.Error(UiString.create())
                 }
             }
+        }
+    }
+
+    private fun getDefaultCategoryId(type: TransactionType): Long {
+        return when (type) {
+            TransactionType.Expense -> DEFAULT_CATEGORY_EXPENSE_ID
+            TransactionType.Income -> DEFAULT_CATEGORY_INCOME_ID
         }
     }
 
@@ -61,9 +72,7 @@ class DatabaseRepositoryImpl @Inject constructor(
         val dbEntity = mapper.mapCategoryEntityToDb(entity)
         return execute {
             val isNameExisted = categoriesDao.isExists(dbEntity.name, dbEntity.type)
-            if (isNameExisted) {
-                throw Exception(CATEGORY_EXISTED)
-            }
+            if (isNameExisted) throw Exception(ERROR_CATEGORY_EXISTED)
             categoriesDao.insert(dbEntity)
         }
     }
@@ -74,13 +83,13 @@ class DatabaseRepositoryImpl @Inject constructor(
     }
 
     override suspend fun deleteCategory(entity: CategoryEntity): ResultWrapper<Int> {
+        if (entity.id == getDefaultCategoryId(entity.type)) {
+            return ResultWrapper.Error(UiString.create(R.string.error_delete_default_category))
+        }
         val dbEntity = mapper.mapCategoryEntityToDb(entity)
         return execute {
             val transactions = transactionsDao.getAllByCategory(dbEntity.id)
-            val newCategoryId = when (TransactionType.valueOf(dbEntity.type)) {
-                TransactionType.Expense -> CATEGORY_EXPENSE_ID
-                TransactionType.Income -> CATEGORY_INCOME_ID
-            }
+            val newCategoryId = getDefaultCategoryId(TransactionType.valueOf(dbEntity.type))
             val mappedTransactions = transactions.map { it.copy(categoryId = newCategoryId) }
             transactionsDao.updateAll(mappedTransactions)
             categoriesDao.delete(dbEntity)
@@ -101,7 +110,9 @@ class DatabaseRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun getCategoriesFlow(type: TransactionType): ResultWrapper<Flow<List<CategoryEntity>>> {
+    override suspend fun getCategoriesFlow(
+        type: TransactionType,
+    ): ResultWrapper<Flow<List<CategoryEntity>>> {
         return execute {
             val flow = categoriesDao.getAllFlow(type.toString())
             flow.map { mapper.mapCategoriesListDbToEntity(it) }
@@ -134,10 +145,13 @@ class DatabaseRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun getTransactionsFlow(): ResultWrapper<Flow<List<TransactionEntity>>> {
+    override suspend fun getTransactionsFlow(
+        type: TransactionType,
+        walletId: Long,
+    ): ResultWrapper<Flow<List<ComplexTransactionEntity>>> {
         return execute {
-            val flow = transactionsDao.getAllFlow()
-            flow.map { mapper.mapTransactionsListDbToEntity(it) }
+            val flow = transactionsDao.getComplexTransactions(type.toString(), walletId)
+            flow.map { mapper.mapComplexTransactionsListDbToEntity(it) }
         }
     }
 
@@ -147,7 +161,11 @@ class DatabaseRepositoryImpl @Inject constructor(
 
     override suspend fun insertCurrency(entity: CurrencyEntity): ResultWrapper<Long> {
         val dbEntity = mapper.mapCurrencyEntityToDb(entity)
-        return execute { currenciesDao.insert(dbEntity) }
+        return execute {
+            val isNameExist = currenciesDao.isExists(dbEntity.name)
+            if (isNameExist) throw Exception(ERROR_CURRENCY_EXISTED)
+            currenciesDao.insert(dbEntity)
+        }
     }
 
     override suspend fun updateCurrency(entity: CurrencyEntity): ResultWrapper<Int> {
@@ -156,14 +174,15 @@ class DatabaseRepositoryImpl @Inject constructor(
     }
 
     override suspend fun deleteCurrency(entity: CurrencyEntity): ResultWrapper<Int> {
+        if (entity.id == prefs.defaultCurrencyId) {
+            return ResultWrapper.Error(UiString.create(R.string.error_delete_default_currency))
+        }
         val dbEntity = mapper.mapCurrencyEntityToDb(entity)
         return execute {
             val transactions = transactionsDao.getAllByCurrency(dbEntity.id)
-            if (transactions.isNotEmpty()) {
-                throw Exception(DELETE_CURRENCY_TRANSACTIONS)
-            }
+            if (transactions.isNotEmpty()) throw Exception(ERROR_DELETE_CURRENCY_WITH_TRANSACTIONS)
             val wallets = walletsDao.getAllByCurrency(dbEntity.id)
-            val mappedWallets = wallets.map { it.copy(currencyId = CURRENCY_ID) }
+            val mappedWallets = wallets.map { it.copy(currencyId = prefs.defaultCurrencyId) }
             walletsDao.updateAll(mappedWallets)
             currenciesDao.delete(dbEntity)
         }
@@ -189,7 +208,11 @@ class DatabaseRepositoryImpl @Inject constructor(
 
     override suspend fun insertWallet(entity: WalletEntity): ResultWrapper<Long> {
         val dbEntity = mapper.mapWalletEntityToDb(entity)
-        return execute { walletsDao.insert(dbEntity) }
+        return execute {
+            val isNameExist = walletsDao.isExists(dbEntity.name)
+            if (isNameExist) throw Exception(ERROR_WALLET_EXISTED)
+            walletsDao.insert(dbEntity)
+        }
     }
 
     override suspend fun updateWallet(entity: WalletEntity): ResultWrapper<Int> {
@@ -198,11 +221,13 @@ class DatabaseRepositoryImpl @Inject constructor(
     }
 
     override suspend fun deleteWallet(entity: WalletEntity): ResultWrapper<Int> {
+        if (entity.id == prefs.defaultWalletId) {
+            return ResultWrapper.Error(UiString.create(R.string.error_delete_default_wallet))
+        }
         val dbEntity = mapper.mapWalletEntityToDb(entity)
         return execute {
             val transactions = transactionsDao.getAllByWallet(dbEntity.id)
-            val mappedTransactions = transactions.map { it.copy(walletId = WALLET_ID) }
-            transactionsDao.updateAll(mappedTransactions)
+            transactions.forEach { transactionsDao.deleteById(it.id) }
             walletsDao.delete(dbEntity)
         }
     }
@@ -221,10 +246,10 @@ class DatabaseRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun getComplexWallet(id: Long): ResultWrapper<ComplexWalletEntity> {
+    override suspend fun getComplexWallets(): ResultWrapper<Flow<List<ComplexWalletEntity>>> {
         return execute {
-            val dbEntity = walletsDao.getComplexWallet(id)
-            mapper.mapComplexWalletDbToEntity(dbEntity)
+            val flow = walletsDao.getComplexWallets()
+            flow.map { mapper.mapComplexWalletsListDbToEntity(it) }
         }
     }
 }
